backend/requirements.txt
# Core Dependencies
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
gevent==24.2.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Authentication
pyjwt==2.8.0
bcrypt==4.1.2

# OpenAI / Azure OpenAI
openai>=1.12.0

# Vector Database
pinecone>=5.0.0

# LlamaParse
llama-parse==0.4.0
llama-index-core==0.10.43  # Pinned for compatibility with protobuf 3.20.3

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Content Generation
python-pptx==0.6.23
Pillow==10.2.0

# Office Files
openpyxl==3.1.2
python-docx==1.1.0
PyPDF2==3.0.1

# Microsoft Graph (OneDrive)
msal==1.24.1

# Video Generation
moviepy==1.0.3
pydub==0.25.1

# Integrations
boxsdk==3.9.2
slack-sdk==3.26.2
PyGithub==2.1.1
notion-client==2.2.1
pyzotero==1.5.18
requests-oauthlib==1.3.1
# Gmail/GDrive - pinned to compatible versions
protobuf==3.20.3
google-auth==2.16.0
google-auth-oauthlib==1.0.0
google-api-core==2.11.0
google-api-python-client==2.88.0
googleapis-common-protos==1.60.0

# Azure Speech (TTS)
azure-cognitiveservices-speech==1.35.0

# AWS
boto3==1.34.19

# Utilities
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
httpx==0.26.0
python-dateutil==2.8.2

# Background Job Processing
celery==5.3.4
redis==5.0.1
kombu==5.3.4

# NLP
spacy==3.7.2
# Note: en_core_web_sm model is downloaded in Dockerfile, not via pip

# Web Scraper - Self-hosted (requests + BeautifulSoup + OCR)
html2text>=2024.2.26
pytesseract>=0.3.10
